# packages/contracts/openapi/v1/public-api.yaml
openapi: 3.1.0
info:
  title: Ecommerce Starter Kit API
  version: 1.0.0
  description: Contrato publico minimo versionado para endpoints base.
servers:
  - url: http://localhost:4000
paths:
  /health:
    get:
      summary: Health check del servicio
      operationId: getHealth
      responses:
        "200":
          description: API saludable
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [ok]
  /version:
    get:
      summary: Version del servicio API
      operationId: getVersion
      responses:
        "200":
          description: Metadata del servicio
          content:
            application/json:
              schema:
                type: object
                required: [name, version, orderStatusExample]
                properties:
                  name:
                    type: string
                  version:
                    type: string
                  orderStatusExample:
                    type: string
                    enum: [pending, paid, shipped, cancelled]
  /products:
    get:
      summary: Lista de productos activos
      operationId: listProducts
      responses:
        "200":
          description: Productos activos visibles en catalogo
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [id, slug, title, priceCents, currency]
                  properties:
                    id:
                      type: string
                    slug:
                      type: string
                    title:
                      type: string
                    priceCents:
                      type: number
                    currency:
                      type: string
                      enum: [EUR, USD]
  /cart:
    post:
      summary: Crear un carrito vacio
      operationId: createCart
      responses:
        "201":
          description: Carrito creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/{cartId}:
    get:
      summary: Obtener un carrito por id
      operationId: getCartById
      parameters:
        - in: path
          name: cartId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Carrito encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "404":
          description: Carrito no encontrado
  /cart/{cartId}/items:
    post:
      summary: Agregar item al carrito
      operationId: addCartItem
      parameters:
        - in: path
          name: cartId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, quantity]
              properties:
                productId:
                  type: string
                quantity:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: Carrito actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/{cartId}/items/{productId}:
    delete:
      summary: Quitar item del carrito
      operationId: removeCartItem
      parameters:
        - in: path
          name: cartId
          required: true
          schema:
            type: string
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Carrito actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /orders:
    post:
      summary: Crear orden desde carrito
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cartId]
              properties:
                cartId:
                  type: string
      responses:
        "201":
          description: Orden creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}:
    get:
      summary: Obtener orden por id
      operationId: getOrderById
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Orden encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Orden no encontrada
  /orders/{orderId}/payment-intent:
    post:
      summary: Crear PaymentIntent de Stripe para una orden pendiente
      operationId: createOrderPaymentIntent
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                idempotencyKey:
                  type: string
      responses:
        "201":
          description: PaymentIntent creado o reutilizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentIntentSession"
        "404":
          description: Orden no encontrada
        "409":
          description: Orden ya pagada o conflicto de idempotencia
        "503":
          description: Stripe no configurado
  /orders/{orderId}/payment/sync:
    post:
      summary: Sincronizar estado de pago consultando Stripe
      operationId: syncOrderPayment
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Orden actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Orden o pago no encontrado
  /webhooks/stripe:
    post:
      summary: Webhook de Stripe (idempotente por event id)
      operationId: stripeWebhook
      responses:
        "200":
          description: Evento recibido
        "400":
          description: Firma invalida
components:
  schemas:
    CartItem:
      type: object
      required:
        [
          productId,
          productSlug,
          productTitle,
          quantity,
          unitPriceCents,
          lineTotalCents,
          currency,
        ]
      properties:
        productId:
          type: string
        productSlug:
          type: string
        productTitle:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPriceCents:
          type: integer
        lineTotalCents:
          type: integer
        currency:
          type: string
          enum: [EUR, USD]
    Cart:
      type: object
      required:
        [
          id,
          status,
          items,
          totalItems,
          totalCents,
          currency,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [active, ordered]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        totalItems:
          type: integer
        totalCents:
          type: integer
        currency:
          type: string
          enum: [EUR, USD]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderItem:
      type: object
      required:
        [
          productId,
          productSlug,
          productTitle,
          quantity,
          unitPriceCents,
          lineTotalCents,
          currency,
        ]
      properties:
        productId:
          type: string
        productSlug:
          type: string
        productTitle:
          type: string
        quantity:
          type: integer
        unitPriceCents:
          type: integer
        lineTotalCents:
          type: integer
        currency:
          type: string
          enum: [EUR, USD]
    PaymentIntentSession:
      type: object
      required: [clientSecret, paymentIntentId, publishableKey]
      properties:
        clientSecret:
          type: string
        paymentIntentId:
          type: string
        publishableKey:
          type: string
    Order:
      type: object
      required:
        [
          id,
          cartId,
          status,
          items,
          totalItems,
          totalCents,
          currency,
          createdAt,
          updatedAt,
          payment,
        ]
      properties:
        id:
          type: string
        cartId:
          type: string
        status:
          type: string
          enum: [pending, paid, shipped, cancelled]
        payment:
          type: object
          nullable: true
          required: [status, stripePaymentIntentId]
          properties:
            status:
              type: string
              enum:
                [
                  requires_payment_method,
                  processing,
                  succeeded,
                  canceled,
                  failed,
                ]
            stripePaymentIntentId:
              type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalItems:
          type: integer
        totalCents:
          type: integer
        currency:
          type: string
          enum: [EUR, USD]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time